refactor: change app structure
