+++ src/resources/app/app.module.ts
@@ -3,9 +3,9 @@ import { ConfigModule } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { UsersModule } from './users/users.module';
-import { User } from './users/user.entity';
-import { AuthModule } from './auth/auth.module';
+import { UsersModule } from '../users/users.module';
+import { User } from '../users/user.entity';
+import { AuthModule } from '../auth/auth.module';
 
 @Module({
   imports: [
